Add timestamp, simplify Memory, change CSS
Change the site header which contains the site title and
slogan so it has the `brand` value for its `id` attribute,
because I'm using `header` elements to contain memory ids
and timestamps.

Add CSS comment about row height.

Add a timestamp to memories.

Delete CSS `h2` style block.

Delete CSS block for `li:before`, since the memory id is being
represented as an `h2` in each memory `header`.

Add rules to style new memory headers (timestamp, memory id).

Add `timestamp` field/column to database/`Memory` class.

Remove `base64_image` as an init argument from `Memory. This
overcomplicates the class, and in reality, we want to force
people to *not* generate base64 and just pass it. The base64
is simply the built-in thumbnailing method. If we want to change
the thumbnailing method for images, it should be a method which
can be overriden, e.g., a thumbnailer that doesn't glitch images.

Delete `from_dict()` method from `Memory`, there will never be
an instance of us wanting to create from a dictionary, it's just
added annoying upkeep.

Update code, e.g., templates, css to reflect timestamp implementation.